Use thread safe wrapper for MVA categories
