+++ RecoEgamma/ElectronIdentification/src/ElectronMVAEstimatorRun2.cc
@@ -16,8 +16,7 @@ ElectronMVAEstimatorRun2::ElectronMVAEstimatorRun2(const edm::ParameterSet& conf
       << "wrong number of category cuts in ElectronMVAEstimatorRun2" << getTag() << std::endl;
 
   for (int i = 0; i < getNCategories(); ++i) {
-      StringCutObjectSelector<reco::GsfElectron> select(categoryCutStrings[i]);
-      categoryFunctions_.push_back(select);
+      categoryFunctions_.emplace_back(categoryCutStrings[i]);
   }
 
   // Initialize GBRForests from weight files
